// korpus-frontend/src/app/models/imenica.ts
export interface Imenica {
    id?: number;
    nomjed: string;
    genjed: string;
    datjed: string;
    akujed: string;
    vokjed: string;
    insjed: string;
    lokjed: string;
    nommno: string;
    genmno: string;
    datmno: string;
    akumno: string;
    vokmno: string;
    insmno: string;
    lokmno: string;
    vrsta?: any;
    vrsta_id?: number;
    varijante?: any[];
    varijantaimenice_set?: any[];
    izmenaimenice_set?: any[];
    vreme?: Date;
    version?: number;
}