// korpus-frontend/src/app/models/glagol.ts
export interface Glagol {
    id?: number;
    infinitiv?: string;
    vid: number;
    rod: number;
    rgp_mj?: string;
    rgp_zj?: string;
    rgp_sj?: string;
    rgp_mm?: string;
    rgp_zm?: string;
    rgp_sm?: string;
    gpp?: string;
    gps?: string;
    oblici?: any[];
    oblikglagola_set?: any[];
    varijante?: any[];
    varijantaglagola_set?: any[];
    izmenaglagola_set?: any[];
    vreme?: Date;
    version?: number;
}